feat: resend verification email
