+++ src/common/components/Auth/index.ts
@@ -2,3 +2,4 @@ export * from "./AuthCard";
 export * from "./LoginForm";
 export * from "./SignupForm";
 export * from "./ResetPasswordForm";
+export * from "./VerificationEmailForm";
